Reorganize: move result table class to framework
